Soft OpenSearch
With an improved mechanism of Query Settings, it is now possible to specify to OpenSearch ENgine to force the assignement of parameters even if they are not described in the OpenSearch Description of the entity to query. The first class in this library to use it is extending the GenericOpenSearchable class with SoftGenericOpenSearchable.

Signed-off-by: Emmanuel Mathot <emmanuel.mathot@gmail.com>